// src/app/page/master/jenis-bunga/jenis-bunga.interface.ts
 import {EventEmitter} from '@angular/core';

    export interface JenisBunga {
	kodeJabatan?: any;
	namaPangkat?: string;
	reportDisplay?: string;
	noUrut?: number;
	kdGolonganPegawai?: number;
	ruang?: string;
	kdPangkatHead?: number;
	KodeExternal?: any;
	kodeExternal?: any;
	namaExternal?: string;
	kdDepartemen?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
  	value?: any;
  	namaGolonganPegawai?: string;
  	id?:any;
  	kode?:any;
  	version?:any;
	namaJabatan?:any;
	jenisJabatan?:any;
	levelJabatan?:any;
	Version?:any;
	namaJenisBunga?: any;
    }